Added more error handeling, optimzed the template

+++ bin/generator.cr
@@ -3,16 +3,12 @@ require "./generator_help"
 class Generator
   def initialize(exercise)
     stuff = GeneratorHelp.new(exercise).to_s
-    exercise_snake = to_snake(exercise)
+    exercise_snake = exercise.gsub("-", "_")
+    raise "Spec directory not found: ./exercises/practice/#{exercise}/spec" unless Dir.exists?("./exercises/practice/#{exercise}/spec")
     file_name = "./exercises/practice/#{exercise}/spec/#{exercise_snake}_spec.cr"
     file = File.new(file_name, mode = "wb")
     File.write(file_name, stuff)
     system("crystal tool format #{file_name}")
-  end
-
-  def to_snake(input)
-    input = input.gsub("-", "_")
-    input
   end
 end
